+++ packages/better-write-app/src/lang/index.ts
@@ -1,5 +1,5 @@
 import { createI18n } from 'vue-i18n'
-import { ptBR, enUS, VueI18nLocales } from 'better-write-localisation'
+import { ptBR, enUS, VueI18nLocales } from 'better-write-languages'
 
 export default createI18n({
   legacy: false,