+++ packages/better-write-app/src/use/factory.ts
@@ -8,6 +8,7 @@ import { useFormat } from './format'
 import { useUtils } from './utils'
 import { usePDFStore } from '@/store/pdf'
 import i18n from '@/lang'
+import { ImageToForcePNG } from 'better-write-image-converter'
 
 export const useFactory = () => {
   const PDF = usePDFStore()
@@ -191,6 +192,7 @@ export const useFactory = () => {
     ) => {
       const _ = document.createElement('input')
       _.type = 'file'
+      _.accept = '.png, .svg, .jpg, .jpeg'
       _.addEventListener('change', function () {
         const file = (this.files as any)[0]
 
@@ -204,13 +206,19 @@ export const useFactory = () => {
           reader.readAsDataURL(file)
         }
 
-        reader.onload = function () {
+        reader.onload = async () => {
           if (utils.support().images(reader)) {
             error && error('bad file')
             return
           }
 
-          const content = entity().create('image', reader.result as string)
+          const raw = await ImageToForcePNG({
+            raw: reader.result as string,
+            width: 2000 as number,
+            height: 2000 as number,
+          })
+
+          const content = entity().create('image', raw)
 
           load && load(content)
         }