+++ packages/better-write-app/vite.config.ts
@@ -9,7 +9,6 @@ import { VitePWA as vitePWA } from 'vite-plugin-pwa'
 import vitePersist from 'vite-plugin-optimize-persist'
 import vitePackageAccess from 'vite-plugin-package-config'
 import vitePackageVersion from 'vite-plugin-package-version'
-import viteChecker from 'vite-plugin-checker'
 import viteFonts from 'vite-plugin-fonts'
 import viteSitemap from 'vite-plugin-pages-sitemap'
 import { viteStdlib } from "./scripts/vite"
@@ -52,7 +51,6 @@ export default defineConfig({
       resolvers: [HeadlessUiResolver()],
     }),
     windiCSS(),
-    viteChecker({ vueTsc: true }),
     viteFonts({
       google: {
         families: [{