+++ packages/better-write-types/src/types/ast.ts
@@ -1,0 +1,14 @@
+export interface TextAST {
+  text: string
+  bold: boolean
+  italic: boolean
+  underline: boolean
+  link: boolean
+}
+
+export interface TextASTElement {
+  text: string
+  bold?: boolean
+  italic?: boolean
+  underline?: boolean
+}